refactor: create task list component
