+++ src/list-inbox.tsx
@@ -1,45 +1,9 @@
-import { Action, ActionPanel, Color, Icon, Keyboard, List, open, showToast, Toast } from "@raycast/api";
+import { Icon, List, showToast, Toast } from "@raycast/api";
 import { useEffect, useState } from "react";
-import { completeTask } from "./lib/api/complete.task";
-import { deleteTask } from "./lib/api/delete-task";
 import { listTasks } from "./lib/api/list-tasks";
+import { TaskList } from "./lib/components/task-list";
 import { OmniFocusTask } from "./lib/types/task";
 import { useValidateRequirements } from "./lib/utils/useValidateRequirements";
-
-function getAccessories(task: OmniFocusTask): List.Item.Accessory[] {
-  const accessories: List.Item.Accessory[] = [];
-  if (task.flagged) {
-    accessories.push({ icon: Icon.Flag });
-  }
-
-  if (task.deferDate) {
-    accessories.push({ tag: { value: new Date(task.deferDate), color: Color.SecondaryText }, tooltip: "Defer until" });
-  }
-
-  if (task.dueDate) {
-    accessories.push({ tag: { value: new Date(task.dueDate), color: Color.Orange }, tooltip: "Due" });
-  }
-
-  if (task.tags.length) {
-    const tagsToShow = task.tags.slice(0, 3);
-    const hasMoreTags = task.tags.length > 3;
-
-    accessories.push(
-      ...tagsToShow.map((t) => ({
-        tag: t,
-      })),
-    );
-
-    if (hasMoreTags) {
-      accessories.push({
-        tag: "...",
-        tooltip: "More tags are associated with this task.",
-      });
-    }
-  }
-
-  return accessories;
-}
 
 export default function ListInboxTasks() {
   const [tasks, setTasks] = useState<OmniFocusTask[]>([]);
@@ -67,38 +31,6 @@ export default function ListInboxTasks() {
     }
   };
 
-  async function actionComplete(id: string) {
-    try {
-      await completeTask(id);
-      await showToast({
-        title: "Task completed!",
-        style: Toast.Style.Success,
-      });
-      await fetchTasks();
-    } catch {
-      await showToast({
-        title: "An error occurred while completing the task.",
-        style: Toast.Style.Failure,
-      });
-    }
-  }
-
-  async function actionDelete(id: string) {
-    try {
-      await deleteTask(id);
-      await showToast({
-        title: "Task deleted!",
-        style: Toast.Style.Success,
-      });
-      await fetchTasks();
-    } catch {
-      await showToast({
-        title: "An error occurred while deleting the task.",
-        style: Toast.Style.Failure,
-      });
-    }
-  }
-
   useEffect(() => {
     if (!loading) {
       if (check) {
@@ -125,41 +57,5 @@ export default function ListInboxTasks() {
     );
   }
 
-  return (
-    <List isLoading={loading || isLoading}>
-      {tasks.length === 0 && <List.EmptyView title="No tasks in inbox" />}
-      {tasks.length > 0 &&
-        tasks.map((t) => {
-          return (
-            <List.Item
-              key={t.id}
-              title={t.name}
-              icon={!t.completed ? Icon.Circle : Icon.Checkmark}
-              accessories={getAccessories(t)}
-              actions={
-                <ActionPanel>
-                  <Action title="Open" onAction={() => open(`omnifocus:///task/${t.id}`)} icon={Icon.Eye} />
-                  <Action
-                    title="Complete"
-                    onAction={async () => {
-                      await actionComplete(t.id);
-                    }}
-                    icon={Icon.CheckCircle}
-                  />
-                  <Action
-                    title="Delete"
-                    style={Action.Style.Destructive}
-                    onAction={async () => {
-                      await actionDelete(t.id);
-                    }}
-                    icon={Icon.Trash}
-                    shortcut={Keyboard.Shortcut.Common.Remove}
-                  />
-                </ActionPanel>
-              }
-            />
-          );
-        })}
-    </List>
-  );
+  return <TaskList isLoading={loading || isLoading} tasks={tasks} />;
 }
